adding frame information to functions table
