+++ x-pack/plugins/profiling/public/components/frame_information_window/get_impact_rows.ts
