+++ x-pack/plugins/profiling/public/components/frame_information_window/get_information_rows.ts
